Process background tasks in separate threads
